+++ sema/type_checker.h
@@ -48,6 +48,8 @@ public:
     void visit(AllocExpr* node) override;
     void visit(LambdaExpr* node) override;
     void visit(AsmStmt* node) override;
+    void visit(UnionDecl* node) override;
+    void visit(SizeofExpr* node) override;
     void visit(ThreadCreateExpr* node) override;
     void visit(ThreadJoinStmt* node) override;
     void visit(ThrowStmt* node) override;